Reply to suggestion:
Getting there! I made a few fixes.

First, about the difference between GF and GFClosure: I agree it's cleaner the way you wrote it, but there is a bug in GenJAX that prevents the use of GFClosures with importance. If this is what stopped me from doing it the GFC way the first time, I should have made a note of it, but at any rate, I have opened GEN-420 for this issue, and we now use GFs throughout except for the user's supplied distributions.

I have changed the call signature for function_family to take one parameter instead of several. This is because we want to preserve the ability to vmap over simulate to produce vectors of functions. Using an invocation like f(*params, x) will blast all the many function's arguments into the arglist, spoiling that ability.

Note that if a function's parameters have a shape like (3,), then a batch sample will have shape (N, 3). The "family" functions want to work with the parameters they understand, which are now the columns of this matrix. I have adopted the convention `a, b = params.T` to extract the parameters with useful names. Then, the vector arithmetic can proceed as usual in JNP style.

That said, for things like pointwise and compose, we use tuples to bring what may be different parameter shapes together, and then destructure them in the helper functions. This makes the exact type of params tricky to state, so I have left some type errors behind, but the inference works and the jacobian wrt parameters seems to work too.